refactor: remove interfaces into .interface; eslint config
