+++ src/components/ModalVideo/index.tsx
@@ -3,7 +3,7 @@
 import { KeyboardEvent, useEffect, useState } from "react";
 import { useTranslations } from "next-intl";
 
-import { Button } from "components/Button";
+import { Button } from "components/ui/Button";
 
 import "styles/colors.scss";
 import styles from "./modalVideo.module.scss";
