+++ src/components/ui/Button/button.module.scss
@@ -1,5 +1,5 @@
-@import "src/styles/colors";
-@import "src/styles/sizes";
+@import "../../../styles/colors";
+@import "../../../styles/sizes";
 
 .customButton {
     height: $height_h56;
