+++ src/components/ui/Button/index.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { Button } from "./index";
+import { Button } from "components/ui/Button/index";
 
 const meta: Meta<typeof Button> = {
     component: Button,
