+++ src/components/ui/Button/index.tsx
@@ -1,18 +1,8 @@
 "use client";
 
-import { MouseEvent, ReactNode } from "react";
+import { CustomButtonProps } from "./button.interface";
 
 import styles from "./button.module.scss";
-
-export interface CustomButtonProps {
-    label?: string;
-    bgcolor?: string;
-    size?: "small" | "medium" | "large" | "full";
-    children?: ReactNode;
-    onClick?: (event: MouseEvent<HTMLButtonElement>) => void;
-    type?: "submit" | "reset" | "button";
-    disabled?: boolean;
-}
 
 export const Button = ({
     label,
@@ -33,7 +23,6 @@ export const Button = ({
             disabled={disabled !== undefined ? disabled : false}
             className={`${styles.customButton} ${buttonSizeClass} ${buttonColorClass} ${buttonDisabledClass}`}
             onClick={onClick}
-            /* eslint-disable react/jsx-props-no-spreading */
             {...props}>
             {label}
             {children}
