+++ src/components/ui/Input/input.interface.ts
@@ -1,0 +1,7 @@
+import { PropsWithRef } from "react";
+import IntrinsicElements = React.JSX.IntrinsicElements;
+
+export interface InputProps extends PropsWithRef<IntrinsicElements["input"]> {
+    className?: string;
+    errorMessage?: string | undefined;
+}
