+++ src/components/ui/Select/select.interface.ts
@@ -1,0 +1,9 @@
+import { SelectHTMLAttributes } from "react";
+
+export interface CustomSelectProps
+    extends SelectHTMLAttributes<HTMLSelectElement> {
+    name: string;
+    label: string;
+    options: { value: string; label: string }[];
+    errorMessage?: string | undefined;
+}
