+++ src/components/ui/TextArea/textarea.interface.ts
@@ -1,0 +1,7 @@
+import { TextareaHTMLAttributes } from "react";
+
+export interface CustomTextAreaProps
+    extends TextareaHTMLAttributes<HTMLTextAreaElement> {
+    className?: string;
+    errorMessage?: string | undefined;
+}
