+++ src/components/ui/TextArea/textarea.module.scss
@@ -1,14 +1,14 @@
-@import "src/styles/sizes";
-@import "src/styles/colors";
+@import "../../../styles/colors";
+@import "../../../styles/sizes";
 
-.input {
-    height: $height_h56;
+.textArea {
+    height: $height_h150;
     width: 100%;
     padding: $padding_p14 0 $padding_p14 $padding_p24;
     margin-right: $margin_m24;
     margin-bottom: $margin_m20;
     background-color: transparent;
-    color: $white;
+    color: $dark-blue;
     border: solid $border_b01 $dark-grey;
     font-size: $fontSize_f16;
 
@@ -17,7 +17,7 @@
     }
 }
 .error {
-    position: relative;
-    top: -$margin_m15;
     color: $red;
+    position: relative;
+    top: -$margin_m32;
 }
